# dcmd: optimize arg parsing