--- a/lib/dcmd/parse.go
+++ b/lib/dcmd/parse.go
@@ -370,12 +370,9 @@ func ParseSwitches(switches []*ArgDef, data *Data, split []*RawArg) ([]*RawArg, 
 	return newRaws, nil
 }
 
-var (
-	ArgContainers = []rune{
-		'"',
-		'`',
-	}
-)
+func isArgContainer(r rune) bool {
+	return r == '"' || r == '`'
+}
 
 type RawArg struct {
 	Str       string
@@ -384,9 +381,9 @@ type RawArg struct {
 
 // SplitArgs splits the string into fields
 func SplitArgs(in string) []*RawArg {
-	rawArgs := make([]*RawArg, 0)
+	var rawArgs []*RawArg
 
-	curBuf := ""
+	var buf strings.Builder
 	escape := false
 	var container rune
 	for _, r := range in {
@@ -394,7 +391,7 @@ func SplitArgs(in string) []*RawArg {
 		if r == '\\' {
 			if escape {
 				escape = false
-				curBuf += "\\"
+				buf.WriteByte('\\')
 			} else {
 				escape = true
 			}
@@ -405,38 +402,31 @@ func SplitArgs(in string) []*RawArg {
 		// Check for other special tokens
 		isSpecialToken := true
 		if r == ' ' {
-			// Maybe seperate by space
-			if curBuf != "" && container == 0 && !escape {
-				rawArgs = append(rawArgs, &RawArg{curBuf, 0})
-				curBuf = ""
-			} else if curBuf != "" {
-				curBuf += " "
+			// Maybe separate by space
+			if buf.Len() > 0 && container == 0 && !escape {
+				rawArgs = append(rawArgs, &RawArg{buf.String(), 0})
+				buf.Reset()
+			} else if buf.Len() > 0 {
+				buf.WriteByte(' ')
 			}
 		} else if r == container && container != 0 {
 			// Split arg here
 			if escape {
-				curBuf += string(r)
+				buf.WriteRune(r)
 			} else {
-				rawArgs = append(rawArgs, &RawArg{curBuf, container})
-				curBuf = ""
+				rawArgs = append(rawArgs, &RawArg{buf.String(), container})
+				buf.Reset()
 				container = 0
 			}
-		} else if container == 0 && curBuf == "" {
+		} else if container == 0 && buf.Len() == 0 {
 			// Check if we should start containing a arg
-			foundMatch := false
-			for _, v := range ArgContainers {
-				if v == r {
-					if escape {
-						curBuf += string(r)
-					} else {
-						container = v
-					}
-					foundMatch = true
-					break
+			if isArgContainer(r) {
+				if escape {
+					buf.WriteRune(r)
+				} else {
+					container = r
 				}
-			}
-
-			if !foundMatch {
+			} else {
 				isSpecialToken = false
 			}
 		} else {
@@ -445,9 +435,9 @@ func SplitArgs(in string) []*RawArg {
 
 		if !isSpecialToken {
 			if escape {
-				curBuf += "\\"
+				buf.WriteByte('\\')
 			}
-			curBuf += string(r)
+			buf.WriteRune(r)
 		}
 
 		// Reset escape mode
@@ -455,11 +445,12 @@ func SplitArgs(in string) []*RawArg {
 	}
 
 	// Something was left in the buffer just add it to the end
-	if curBuf != "" {
+	if buf.Len() > 0 {
+		item := buf.String()
 		if container != 0 {
-			curBuf = string(container) + curBuf
+			item = string(container) + item
 		}
-		rawArgs = append(rawArgs, &RawArg{curBuf, 0})
+		rawArgs = append(rawArgs, &RawArg{item, 0})
 	}
 
 	return rawArgs